Move README.md file and updated installation script for the number guessing game.
The README.md file was moved from guessing_game/number_guessing_game/ to guessing_game/. This file now provides an overview of the Number Guessing Game project, explaining that it is a simple game where users try to guess a randomly generated number.

In the installation script, the path for generating the README.md file was updated to reflect this change.